Refactoring unit tests: separated DatabaseGenerator build and dbupgrade.
