+++ Source/Rhetos.DatabaseGenerator.Test/Rhetos.DatabaseGenerator.Test.csproj
@@ -62,11 +62,13 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="ConceptApplicationRepositoryTest.cs" />
+    <Compile Include="DatabaseModelBuilderTest.cs" />
     <Compile Include="DatabaseGeneratorTest.cs" />
     <Compile Include="DatabaseModelBuilderAccessor.cs" />
     <Compile Include="DataMigrationScriptBuilderTest.cs" />
     <Compile Include="MockConceptApplicationRepository.cs" />
     <Compile Include="MockDatabaseModel.cs" />
+    <Compile Include="MockDatabasePluginsContainer.cs" />
     <Compile Include="MockDslModel.cs" />
     <Compile Include="MockSqlExecuter.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
